[fix][recurrent-block] preserve h.dtype through ACT accumulator under bf16/fp16
bench_vs_transformer.py at --dtype bf16 crashed in OpenMythos.forward
with 'expected mat1 and mat2 to have the same dtype, but got: float
!= c10::BFloat16' at GQAttention.wq inside the coda. Root cause is
in RecurrentBlock.forward: the ACT halting accumulator silently
upcasts h_out to fp32, which then breaks downstream coda layers that
hold bf16 weights.

Failure trace (bf16):

  RecurrentBlock.forward:
    cumulative_p = torch.zeros(...)                  # NO dtype -> fp32
    weight = torch.where(cumulative_p + p >= ..., remainder, p)
                                                     # remainder fp32 + p bf16
                                                     # torch.where -> fp32
    h_out = h_out + weight.unsqueeze(-1) * h         # fp32 * bf16 -> fp32
                                                     # h_out becomes fp32
  ...returns fp32 h_out to coda which has bf16 wq -> crash

Fixes by:

  * Making cumulative_p explicitly fp32 with a comment explaining why
    (sigmoid-sum accumulator over up to n_loops iterations needs the
    precision headroom for a clean compare against act_threshold ~ 0.99).
    No-op for fp32 callers; explicit for bf16/fp16 callers.
  * Casting weight back to h.dtype right before the h_out += weight * h
    step so the accumulator stays in h's dtype. This is the actual fix
    that unbreaks bf16/fp16 inference. fp32 callers see a no-op cast.
  * Promoting p to fp32 (.float()) for the accumulator arithmetic.
    Slight precision improvement for bf16/fp16; no-op for fp32.

Verified by re-running tests/bench_vs_transformer.py at --size 1b and
--size 3b with --dtype bf16 on rtx6000 GPU 1 (Blackwell PRO 6000,
driver 610.43.02, fallback SDPA path because flash-attn is not yet
installed). Both complete cleanly:

  1b bf16:  Baseline 594ms, OpenMythos loops=1 80ms, peak 11.1 GB
  3b bf16:  Baseline 651ms, OpenMythos loops=1 88ms, peak 32.5 GB

Scientific behavior preserved: fp32 training runs were unaffected
(every change is a no-op when h.dtype == fp32). bf16/fp16 inference
is now correct rather than silently broken.

+++ open_mythos/main.py
@@ -852,7 +852,18 @@ class RecurrentBlock(nn.Module):
         B, T, D = h.shape
 
         halted = torch.zeros(B, T, device=h.device, dtype=torch.bool)
-        cumulative_p = torch.zeros(B, T, device=h.device)
+        # cumulative_p is the ACT halting-probability accumulator. Kept in
+        # fp32 regardless of h.dtype because:
+        #   * sums of up to n_loops sigmoid values need precision headroom to
+        #     compare reliably against act_threshold (typically ~0.99)
+        #   * p comes from sigmoid(Linear(h)) and is promoted to fp32 here
+        #     anyway via the addition; making the storage fp32 keeps that
+        #     explicit and stable across loop iterations
+        # The fp32 weight is cast back to h.dtype right before scaling h on
+        # line 881 so h_out preserves h's dtype (was a bug: with h in bf16,
+        # weight stayed fp32 and the bf16*fp32 promote made h_out fp32,
+        # which then failed in the coda's bf16 attention weights).
+        cumulative_p = torch.zeros(B, T, device=h.device, dtype=torch.float32)
         h_out = torch.zeros_like(h)
 
         for t in range(n_loops):
@@ -863,7 +874,7 @@ class RecurrentBlock(nn.Module):
             trans_out = trans_out + self.lora(trans_out, t)
             h = self.injection(h, e, trans_out)
 
-            p = self.act(h)  # (B, T)
+            p = self.act(h).float()  # (B, T) -- accumulate in fp32 regardless of h.dtype
             still_running = ~halted
 
             # ACT remainder trick: once cumulative_p + p crosses threshold,
@@ -878,7 +889,11 @@ class RecurrentBlock(nn.Module):
                 p,
             )
             weight = weight * still_running.float()
-            h_out = h_out + weight.unsqueeze(-1) * h
+            # Cast weight back to h.dtype before scaling h so the accumulator
+            # stays in h's dtype. Without this cast, bf16/fp16 forwards
+            # silently upcast h_out to fp32 and break downstream coda layers
+            # that expect bf16 inputs against bf16 weights.
+            h_out = h_out + weight.unsqueeze(-1).to(h.dtype) * h
 
             cumulative_p = cumulative_p + p * still_running.float()
             halted = halted | (cumulative_p >= self.cfg.act_threshold)
